refactor(cron): split service operations
